+++ changelog.md
@@ -2,6 +2,17 @@
 
 > All Time Stamps: GMT+8
 
+## 0.2.43.0 _2020-02-04 10:30_
+
+* 增加方块和物品 added blocks and items
+  * 葡萄石矿 prehnite ore block
+  * 葡萄石 prehnite
+  * 葡萄石块 prehnite
+* 增加材料 added materials
+  * 葡萄石 prehnite
+* 增加属性 added traits
+  * 生命激发 life inspiring
+
 ## 0.2.42.0 _2020-02-03 10:00_
 
 * 增加方块和物品 added blocks and items